Point is basically just a pair, with x as first, and y as second
after we'd made SPoint immutable the only difference between it and Pair were the property names and the added function for multiplication with Matrix.
well, both of those can be added as extensions instead, so we'll do that. (specific to the type alias of Pair<Double, Double>, so x and y don't just show up for any arbitrary pair.)