Add an indexer property to S2DMatrix, really access it as a 2D array
the values are actually stored in a hidden 1D array called mm, but the 2D get/set does all the index translation.
we're also going with the traditional one-based indexes, so the matrix is accessed like [1..3, 1..3].  makes matrix multiplication so much easier to write as the usual algorithm with fewer opportunities for typos to introduce error.
 the index translation is done by a single internal function (called 'i') to ensure get/set always derive the 1D index identically.