NeuralNet using asList/subList (on array) instead of classes
the combination of asList and subList creates 'views' into DoubleArray (the equivalent of a Java array of primitive doubles)
It seems equivalent to using ReadOnlySpan in dotnet, protecting the data from being changed while reducing boxing.
As the genetic algorithm works with an array of weights, we get the benefit of the NeuralNet not having to flatten Layer/Neuron to look similar.  NeuralNet's update remains able to calculate by folding outputs of the previous layer as inputs to the next, keeping the algorithm as-is.